+++ src/LL/LLSharp.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode';
 import regex2diagnostics from '../util/regex2diagnostics';
 
-export default function LLSharp(doc: vscode.TextDocument): vscode.Diagnostic[] {
+export default function LLSharp(doc: vscode.TextDocument, txt: string): vscode.Diagnostic[] {
     return regex2diagnostics(
-        doc,
+        doc, txt,
         "LLSharp",
         /\\sharp/g,
     );