+++ src/util/regex2diagnostics.ts
@@ -2,9 +2,9 @@ import * as vscode from 'vscode';
 import { LLCode, messages } from './constants';
 import ranges2diagnostics from './ranges2diagnostics';
 
-export default function regex2diagnostics(doc: vscode.TextDocument, code: LLCode, pattern: RegExp): vscode.Diagnostic[] {
+export default function regex2diagnostics(doc: vscode.TextDocument, txt: string, code: LLCode, pattern: RegExp): vscode.Diagnostic[] {
     const ranges: vscode.Range[] = [];
-    for (const match of doc.getText().matchAll(pattern)) {
+    for (const match of txt.matchAll(pattern)) {
         const startPos = doc.positionAt(match.index);
         const endPos = startPos.translate(0, match[0].length);
         ranges.push(new vscode.Range(startPos, endPos));